requirements.txt
aiosqlite==0.21.0
discord==2.6.3
python-dotenv==1.1.1